+++ artemis-maven/src/main/java/com/artemis/ArtemisMaven.java
@@ -10,6 +10,7 @@ import java.util.List;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
+import org.apache.maven.plugin.logging.Log;
 import org.apache.maven.plugins.annotations.Component;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
@@ -18,7 +19,8 @@ import org.sonatype.plexus.build.incremental.BuildContext;
 import com.artemis.meta.ClassMetadata;
 
 /**
- * The artemis plugin does bytecode-weaving on annotated components.
+ * The artemis plugin does bytecode-weaving on annotated components
+ * and related classes.
  */
 @Mojo(name="artemis", defaultPhase=PROCESS_CLASSES)
 public class ArtemisMaven extends AbstractMojo {
@@ -26,38 +28,51 @@ public class ArtemisMaven extends AbstractMojo {
 	/**
 	 * Root folder for class files.
 	 */
-	@Parameter(property="project.build.outputDirectory")
+	@Parameter(property="project.build.outputDirectory", readonly=true)
 	private File outputDirectory;
 
 	/**
 	 * Root source folder.
 	 */
-	@Parameter(property="project.build.sourceDirectory")
+	@Parameter(property="project.build.sourceDirectory", readonly=true)
 	private File sourceDirectory;
 	
 	/**
 	 * If true, will leave field stubs to keep IDE:s happy after transformations.
 	 */
-	@Parameter
+	@Parameter(property="ideFriendlyPacking")
 	private boolean ideFriendlyPacking;
 	
 	/**
 	 * Enabled weaving of pooled components (more viable on Android than JVM).
 	 */
-	@Parameter(defaultValue="true")
+	@Parameter(defaultValue="true", property="enablePooledWeaving")
 	private boolean enablePooledWeaving;
+	
+	/**
+	 * If false, no weaving will take place (useful for debugging).
+	 */
+	@Parameter(defaultValue="true", property="enableArtemisPlugin")
+	private boolean enableArtemisPlugin;
 	
 	@Component
 	private BuildContext context;
 	
 	@Override
 	public void execute() throws MojoExecutionException, MojoFailureException {
+		if (!enableArtemisPlugin) {
+			getLog().info("Plugin disabled via 'enableArtemisPlugin' set to false.");
+			return;
+		}
+		
 		long start = System.currentTimeMillis();
 		if (context != null && !context.hasDelta(sourceDirectory))
 			return;
 		
-		getLog().info(String.format("Configuration:\n\tideFriendlyPacking=%b \n\tenablePooledWeaving=%b ", 
-			ideFriendlyPacking, enablePooledWeaving));
+		Log log = getLog();
+		log.info("Configuration:"); 
+		log.info("\t ideFriendlyPacking=" + ideFriendlyPacking);
+		log.info("\tenablePooledWeaving=" + enablePooledWeaving);
 		
 		Weaver.retainFieldsWhenPacking(ideFriendlyPacking);
 		Weaver.enablePooledWeaving(enablePooledWeaving);
@@ -65,13 +80,12 @@ public class ArtemisMaven extends AbstractMojo {
 		Weaver weaver = new Weaver(outputDirectory);
 		List<ClassMetadata> processed = weaver.execute();
 		
-		getLog().info(getSummary(processed, start));
+		log.info(getSummary(processed, start));
 	}
 	
 	private static CharSequence getSummary(List<ClassMetadata> processed, long start) {
 		int pooled = 0, packed = 0;
-		for (ClassMetadata meta : processed)
-		{
+		for (ClassMetadata meta : processed) {
 			if (PACKED == meta.annotation) packed++;
 			else if (POOLED == meta.annotation) pooled++;
 		}